AI/ai/requirements.txt
rclpy>=0.1.0
std_msgs>=0.1.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
paho-mqtt>=1.6.1 